First commit
Adding lecture slides for the first 3 lectures and solution for the first assignment -
practice with key-value stores. The database we had to use was Redis and the main task
was to use it's transaction mechanism. Implemented a toy crypto exchange, where the
transaction is to buy cryptocurrency for a fixed price. If the time for the fixed price
comes to an end before making a purchase, the transaction is not committed to the database.