Add Cassandra exercise solution
Adding solution to the 3rd exercise, designed to get hands-on experience
with wide-column stores. We had to model a database suitable for this model
and write a simple program using the created database.

Modelled and implemented a service, which allows to store and read
sensor data from wind turbines in a wind farm.