Add Neo4j exercise solutions
Adding solution to the 4th exercise, designed to get hands-on experience
with graph databases. We had to model a database suitable for this model
and write a simple program to use the created database.

Modelled and implemented a database for a fictional game - Thief,
which allows to get information about player - it's friends, ability
points, bought and unlocked abilities and also shortest path and cost
to unlock a certain ability.